pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "altair[all]>=5.5.0",
  "biopython>=1.85",
  "boto3>=1.38.33",
  "dnachisel>=3.2.16",
  "msgspec>=0.19.0",
  "plotnine>=0.14.5",
  "polars>=1.30.0",
  "viennarna>=2.7.0",
]

[dependency-groups]
dev = [
  "ipython>=9.2.0",
  "jupyterlab>=4.4.3",
  "pyright>=1.1.401",
  "pytest>=8.3.5",
  "ruff>=0.11.12",
]

[tool.pytest.ini_options]
addopts = "--doctest-modules"
